requirements.txt
Flask~=2.0.2
flask-restx~=0.5.1
Flask-CORS~=3.0.10
Flask-SQLAlchemy~=2.5.1
marshmallow~=3.14.1
PyJWT~=2.3.0
gunicorn~=20.1.0
psycopg2~=2.9.2
pytest
pytest-cov
freezegun
flake8
mypy
black